ci: add .gitlab-ci.yml
This is not the ideal setup, but it's a start.

I'd rather not have the tests split out like this but when they aren't,
everything after the first test fails with `mounting /dev/vda on /mnt-root/
failed: No such file or directory`.  I don't know what that's about.  Check it
out here: https://gitlab.com/eqyiel/nixos-mailserver/-/jobs/90278230

These also failed when I tested them, but for a different reason (timeouts).
@griff has made a PR to increase the DefaultTimeoutStartSec for tests:
https://github.com/NixOS/nixpkgs/pull/44916